Move app invite Activity to invite package
