+++ app/src/main/AndroidManifest.xml
@@ -151,8 +151,9 @@
       </intent-filter>
     </activity>
     <activity
-      android:name=".activity.AppInviteDeepLinkActivity"
-      android:theme="@style/Theme.GDG.Panel"></activity>
+      android:name=".invite.AppInviteDeepLinkActivity"
+      android:theme="@style/Theme.GDG.Panel">
+    </activity>
     <activity
       android:name=".activity.SearchActivity"
       android:launchMode="singleTop">
